+++ sdk/core/azure-core/azure/core/tracing/__init__.py
@@ -2,12 +2,7 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT License.
 # ------------------------------------
-from azure.core.tracing._abstract_span import (
-    AbstractSpan,
-    SpanKind,
-    HttpSpanMixin,
-)
+from ._abstract_span import AbstractSpan, HttpSpanMixin
 from ._models import Link, SpanKind, StatusCode, TracingOptions
-from ._tracer import TracerProvider
 
-__all__ = ["AbstractSpan", "HttpSpanMixin", "Link", "SpanKind", "StatusCode", "TracingOptions", "TracerProvider"]
+__all__ = ["AbstractSpan", "HttpSpanMixin", "Link", "SpanKind", "StatusCode", "TracingOptions"]