+++ sdk/core/azure-core/azure/core/tracing/decorator_async.py
@@ -27,20 +27,16 @@
 
 import functools
 
-from typing import (
-    Awaitable,
-    Callable,
-    Any,
-    TypeVar,
-    overload,
-    Optional,
-    Mapping,
-)
+from typing import Awaitable, Callable, Any, TypeVar, overload, Optional, Mapping, TYPE_CHECKING
 from typing_extensions import ParamSpec
 from .common import change_context, get_function_and_class_name
-from ._models import TracingOptions, SpanKind
-from ._tracer import default_tracer_provider, TracerProvider
+from ._models import SpanKind as _SpanKind
+from ..instrumentation import default_instrumentation
 from ..settings import settings
+
+if TYPE_CHECKING:
+    from ._models import TracingOptions
+    from ..instrumentation import Instrumentation
 
 
 P = ParamSpec("P")
@@ -56,9 +52,9 @@ def distributed_trace_async(__func: Callable[P, Awaitable[T]]) -> Callable[P, Aw
 def distributed_trace_async(
     *,
     name_of_span: Optional[str] = None,
-    kind: Optional["SpanKind"] = None,
+    kind: Optional[_SpanKind] = None,
     tracing_attributes: Optional[Mapping[str, Any]] = None,
-    tracer_provider: Optional[TracerProvider] = None,
+    instrumentation: Optional["Instrumentation"] = None,
     **kwargs: Any,
 ) -> Callable[[Callable[P, Awaitable[T]]], Callable[P, Awaitable[T]]]:
     pass
@@ -68,9 +64,9 @@ def distributed_trace_async(  # pylint: disable=unused-argument
     __func: Optional[Callable[P, Awaitable[T]]] = None,
     *,
     name_of_span: Optional[str] = None,
-    kind: Optional["SpanKind"] = None,
+    kind: Optional[_SpanKind] = None,
     tracing_attributes: Optional[Mapping[str, Any]] = None,
-    tracer_provider: Optional[TracerProvider] = None,
+    instrumentation: Optional["Instrumentation"] = None,
     **kwargs: Any,
 ) -> Any:
     """Decorator to apply to function to get traced automatically.
@@ -92,15 +88,15 @@ def distributed_trace_async(  # pylint: disable=unused-argument
     :paramtype kind: ~azure.core.tracing.SpanKind
     :keyword tracing_attributes: Attributes to add to the span.
     :paramtype tracing_attributes: Mapping[str, Any] or None
-    :keyword tracer_provider: The tracer provider to use. If not provided, a default tracer provider will be used.
-    :paramtype tracer_provider: ~azure.core.tracing.TracerProvider
+    :keyword instrumentation: The instrumentation instance to use. If not provided, a default one will be used.
+    :paramtype instrumentation: ~azure.core.instrumentation.Instrumentation
     :return: The decorated function
     :rtype: Any
     """
     if tracing_attributes is None:
         tracing_attributes = {}
     if kind is None:
-        kind = SpanKind.INTERNAL
+        kind = _SpanKind.INTERNAL
 
     def decorator(func: Callable[P, Awaitable[T]]) -> Callable[P, Awaitable[T]]:
         @functools.wraps(func)
@@ -133,6 +129,7 @@ def distributed_trace_async(  # pylint: disable=unused-argument
 
             name = name_of_span or func.__qualname__
             if span_impl_type:
+                # Plugin path
                 with change_context(passed_in_parent):
                     name = name_of_span or get_function_and_class_name(func, *args)
                     with span_impl_type(name=name, kind=kind) as span:
@@ -140,8 +137,9 @@ def distributed_trace_async(  # pylint: disable=unused-argument
                             span.add_attribute(key, value)  # type: ignore
                         return await func(*args, **kwargs)
             else:
+                # Native path
                 method_tracer = (
-                    tracer_provider.get_tracer() if tracer_provider else default_tracer_provider.get_tracer()
+                    instrumentation.get_tracer() if instrumentation else default_instrumentation.get_tracer()
                 )
                 if not method_tracer:
                     return await func(*args, **kwargs)
@@ -149,7 +147,6 @@ def distributed_trace_async(  # pylint: disable=unused-argument
                     name=name,
                     kind=kind,
                     attributes=span_attributes,
-                    record_exception=tracing_options.get("record_exception", True),
                 ):
                     return await func(*args, **kwargs)