+++ sdk/core/azure-core/tests/async_tests/test_tracing_decorator_async.py
@@ -16,15 +16,15 @@ from azure.core.pipeline import Pipeline, PipelineResponse
 from azure.core.pipeline.policies import HTTPPolicy
 from azure.core.pipeline.transport import HttpTransport
 from azure.core.settings import settings
-from azure.core.tracing import SpanKind, TracerProvider
-from azure.core.tracing.decorator import distributed_trace
+from azure.core.instrumentation import Instrumentation
+from azure.core.tracing import SpanKind
 from azure.core.tracing.decorator_async import distributed_trace_async
-from opentelemetry.trace import SpanKind as OtelSpanKind, StatusCode as OtelStatusCode
+from opentelemetry.trace import StatusCode as OtelStatusCode
 
 from tracing_common import FakeSpan
 from utils import HTTP_REQUESTS
 
-custom_tracer_provider = TracerProvider(
+custom_instrumentation = Instrumentation(
     library_name="my-library",
     library_version="1.0.0",
     schema_url="https://test.schema",
@@ -92,8 +92,8 @@ class MockClient:
     async def raising_exception(self, **kwargs):
         raise ValueError("Something went horribly wrong here")
 
-    @distributed_trace_async(tracer_provider=custom_tracer_provider)
-    async def method_with_custom_tracer_provider(self):
+    @distributed_trace_async(instrumentation=custom_instrumentation)
+    async def method_with_custom_instrumentation(self):
         time.sleep(0.001)
 
     @distributed_trace_async
@@ -265,16 +265,16 @@ class TestAsyncDecoratorNativeTracing:
 
     @pytest.mark.asyncio
     @pytest.mark.parametrize("http_request", HTTP_REQUESTS)
-    async def test_decorator_custom_tracer_provider(self, tracing_helper, http_request):
+    async def test_decorator_custom_instrumentation(self, tracing_helper, http_request):
         """Test that a custom tracer provider can be used."""
         client = MockClient(http_request)
         settings.tracing_enabled = True
         with tracing_helper.tracer.start_as_current_span("Root"):
-            await client.method_with_custom_tracer_provider()
+            await client.method_with_custom_instrumentation()
 
             finished_spans = tracing_helper.exporter.get_finished_spans()
             assert len(finished_spans) == 1
-            assert finished_spans[0].name == "MockClient.method_with_custom_tracer_provider"
+            assert finished_spans[0].name == "MockClient.method_with_custom_instrumentation"
             assert finished_spans[0].instrumentation_scope.schema_url == "https://test.schema"
             assert finished_spans[0].instrumentation_scope.name == "my-library"
             assert finished_spans[0].instrumentation_scope.version == "1.0.0"
@@ -321,23 +321,3 @@ class TestAsyncDecoratorNativeTracing:
 
             finished_spans = tracing_helper.exporter.get_finished_spans()
             assert len(finished_spans) == 0
-
-    @pytest.mark.asyncio
-    @pytest.mark.parametrize("http_request", HTTP_REQUESTS)
-    async def test_decorated_method_without_recording_exception(self, tracing_helper, http_request):
-        """Test that exceptions being recorded as events can be suppressed."""
-        client = MockClient(http_request)
-        settings.tracing_enabled = True
-        with tracing_helper.tracer.start_as_current_span("Root"):
-            try:
-                await client.raising_exception(tracing_options={"record_exception": False})
-            except ValueError:
-                pass
-
-            finished_spans = tracing_helper.exporter.get_finished_spans()
-            assert len(finished_spans) == 1
-            assert finished_spans[0].name == "MockClient.raising_exception"
-            assert finished_spans[0].status.status_code == OtelStatusCode.ERROR
-            assert "Something went horribly wrong here" in finished_spans[0].status.description
-            assert finished_spans[0].attributes.get("error.type") == "ValueError"
-            assert not finished_spans[0].events